+++ apps/ledger-live-desktop/src/renderer/drawers/DataSelector/SelectCurrencyDrawer.jsx
@@ -1,24 +1,23 @@
 // @flow
 
-import React, { useMemo, useState, memo, useCallback } from "react";
+import React, { useMemo, useState, memo } from "react";
 import styled from "styled-components";
 import { useTranslation } from "react-i18next";
 import { SearchInput } from "@ledgerhq/react-ui";
-import { Account, AccountLike, Currency } from "@ledgerhq/live-common/lib/types";
-import { setDrawer } from "../Provider";
 
 import Fuse from "fuse.js";
+
+import type { Currency } from "@ledgerhq/live-common/types/index";
 
 import {
   listSupportedCurrencies,
   listTokens,
   useCurrenciesByMarketcap,
-} from "@ledgerhq/live-common/lib/currencies";
-import { makeRe } from "@ledgerhq/live-common/lib/platform/filters";
+} from "@ledgerhq/live-common/currencies/index";
+import { makeRe } from "@ledgerhq/live-common/platform/filters";
 import Text from "~/renderer/components/Text";
-import { ThemedComponent } from "~/renderer/styles/StyleProvider";
+import type { ThemedComponent } from "~/renderer/styles/StyleProvider";
 import { CurrencyList } from "./CurrencyList";
-import SelectAccountDrawer from "./SelectAccountDrawer";
 
 const options = {
   includeScore: false,
@@ -63,10 +62,9 @@ const HeaderContainer: ThemedComponent<any> = styled.div`
 `;
 
 type SelectCurrencyDrawerProps = {
-  onClose: () => void,
   currencies?: string[],
   includeTokens?: boolean,
-  onAccountSelected: (account: AccountLike, parentAccount?: Account) => void,
+  onCurrencySelected: (currency: Currency) => void,
 };
 
 const SearchInputContainer = styled.div`
@@ -74,9 +72,11 @@ const SearchInputContainer = styled.div`
   flex: 0 1 auto;
 `;
 
-const SelectCurrencyDrawer = (props: SelectCurrencyDrawerProps) => {
-  const { currencies, includeTokens, onAccountSelected, onClose } = props;
-
+const SelectCurrencyDrawer = ({
+  currencies,
+  includeTokens,
+  onCurrencySelected,
+}: SelectCurrencyDrawerProps) => {
   const { t } = useTranslation();
 
   const cryptoCurrencies = useMemo(() => {
@@ -113,38 +113,6 @@ const SelectCurrencyDrawer = (props: SelectCurrencyDrawerProps) => {
     return fuzzySearch(sortedCurrencies, searchValue);
   }, [searchValue, sortedCurrencies]);
 
-  const handleCurrencySelected = useCallback(
-    currency => {
-      setDrawer(
-        SelectAccountDrawer,
-        {
-          currency,
-          onAccountSelected,
-          onRequestBack: () =>
-            setDrawer(SelectCurrencyDrawer, props, {
-              onRequestClose: () => {
-                console.log("REQUEST CLOSE");
-                onClose();
-              },
-            }),
-        },
-        {
-          onRequestClose: () => {
-            console.log("REQUEST CLOSE");
-            onClose();
-          },
-        },
-      );
-    },
-    [onAccountSelected, props, onClose],
-  );
-
-  if (cryptoCurrencies.length === 1) {
-    return (
-      <SelectAccountDrawer currency={cryptoCurrencies[0]} onAccountSelected={onAccountSelected} />
-    );
-  }
-
   return (
     <SelectCurrencyDrawerContainer>
       <HeaderContainer>
@@ -161,7 +129,7 @@ const SelectCurrencyDrawer = (props: SelectCurrencyDrawerProps) => {
         <SearchInputContainer>
           <SearchInput value={searchValue} onChange={setSearchValue} />
         </SearchInputContainer>
-        <CurrencyList currencies={filteredCurrencies} onCurrencySelect={handleCurrencySelected} />
+        <CurrencyList currencies={filteredCurrencies} onCurrencySelect={onCurrencySelected} />
       </SelectorContent>
     </SelectCurrencyDrawerContainer>
   );
